Add files via upload
Background

Hypothesis testing is a method for evaluating an assumption about a population based on sample data. In the statistics courses I teach, we cover both parametric and non-parametric methods for hypothesis testing. I have observed that students often struggle to select the most appropriate test. At times, it seems as though the primary objective of statistical classes is to teach how to choose the correct test, which is not entirely accurate.

Hypothesis testing helps us determine whether the differences between two or more groups are real or merely due to random variation. While traditional statistical tests such as the t-test, Mann-Whitney U test, and Chi-square test are widely used, we can also apply permutation tests as an alternative.

In a permutation test, when comparing two groups (e.g., Group A and Group B), the entire data set is combined, and data points are re sampled into Groups A and B repeatedly—typically 10,000 or more times. Relevant statistics are calculated for each iteration, and the p-value is then computed, representing the proportion of times the re sampled statistics exceed the observed difference. Notably, permutation tests do not rely on underlying assumptions about the data's distribution.

In contrast, traditional statistical tests require the calculation of test statistics based on predefined distributions, such as the t-distribution or Chi-square distribution, to derive the p-value. These methods involve specific assumptions about the data (e.g., normality, equal variances), which guide the selection of the appropriate test.

Goal and the Data set

As the title suggests, this study focuses on comparing traditional statistical tests with permutation tests. For the traditional statistical tests, we will select the appropriate method based on the type of variables and whether the assumptions of normality are met. In contrast, we will also perform permutation tests to examine their effectiveness.

The dataset used in this analysis is a survey conducted on 600 students who took undergraduate admission tests in Bangladesh. The data was sourced from Kaggle. A separate project on exploratory data analysis (EDA) has already been conducted, and you can refer to that work for a detailed understanding of the dataset.

The dataset contains 15 variables, with University—indicating whether a student was admitted to private or public universities—being the target variable. Among the remaining 14 variables, HSC GPA and SSC GPA are numerical, while the rest are categorical.